FROMLIST: drm/rockchip: Use an enum to identify Rockchip VOPs
We used to call drm_of_encoder_active_endpoint_id() from
rockchip_dp_drm_encoder_atomic_check() to determine the endpoint for the
active encoder. However, the encoder isn't necessarily active at this
point or it may be connected to different crtc than what we're switching
to. Instead, look at the crtc from the drm_crtc_state.  This fixes wrong
colors when driving the eDP with the big VOP.  Further, we can identify
the type of VOP we're dealing with by just putting a VOP id enum in the
vop_data.

On way to test this is to use the modetest tool from libdrm:

  $ modetest -M rockchip -s 32@28:2400x1600

which displays dark or black colors because we fail to look up the
endpoint and use default 0 (which is ROCKCHIP_OUT_MODE_P888) for big
VOP instead of RGB10 as required.

For reference,

  $ modetest -M rockchip -s 32@25:2400x1600

drives the eDP from little VOP and displays correctly.

Link: https://patchwork.freedesktop.org/patch/119573/

BUG=chrome-os-partner:56407
TEST=verify 'modetest -M rockchip -s 32@28:2400x1600' looks right
Signed-off-by: Kristian H. Kristensen <hoegsberg@chromium.org>
Change-Id: If5c5f36bcee09113008ee5155a13337f0e69371f
Reviewed-on: https://chromium-review.googlesource.com/404908
Reviewed-by: Sean Paul <seanpaul@chromium.org>