CHROMIUM: Revert "UPSTREAM: ASoC: Intel: Load the atom DPCM driver only"
This reverts commit ea2b14152cf0 ("UPSTREAM: ASoC: Intel: Load the atom
DPCM driver only").

Reason being I can't select the baytrail audio driver for max98090
codec (what we use) because Intel left no way to enable it *IF*
support for any other codecs is enabled. This was noticed but was
snubbed by Intel:

https://groups.google.com/forum/m/#!topic/linux.kernel/6iP9F-w_yeM
http://lkml.iu.edu/hypermail/linux/kernel/1608.1/03537.html

Redhat bug on the same issue was closed as "Can't fix":
https://bugzilla.redhat.com/show_bug.cgi?id=1335196
("Unfortunately it looks like the decision was made to not support this
module in a distro like environment. I'm going to close
this as CANTFIX unless the Intel developers decide to change their mind
or someone reverts it.")

BUG=chrome-os-partner:59912
TEST=Check that audio works even after reverting this commit in reef

Change-Id: Ie7fcff2c96377e67a307768eec714288e50c1519
Reviewed-on: https://chromium-review.googlesource.com/413626
Commit-Ready: Rajat Jain <rajatja@chromium.org>
Tested-by: Rajat Jain <rajatja@chromium.org>
Reviewed-by: Rajat Jain <rajatja@chromium.org>