CHROMIUM: config: turn off CMA on arm64
There's no known reason why we have CMA turned on for arm64.  It looks
as if we copied it over from the arm64 config from 3.18 without a
second thought and just left it on.  For all arm64 configs in kernel
4.4 we've got iommus and those are used when we want a contiguous
chunk of memory for graphics or video.

Let's get rid of a bunch of extra code and complexity in the kernel by
turning off CMA.

BUG=chrome-os-partner:60243
TEST=for i in $(seq 100); do ./launchBalloons.sh 6 700 30 >/dev/null; done

Change-Id: I637475f947323af0c4fbae02f7dd4b5fb8e817e2
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/414433
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Pawel Osciak <posciak@chromium.org>