CHROMIUM: tpm: cr50_i2c: Disable building as module
Attempts to build the CR50 I2C driver as module fails with

ERROR: "tpm_transmit_cmd" [drivers/char/tpm/cr50_i2c.ko] undefined!

This is seen because tpm_transmit_cmd() is not exported. Instead of
exporting a symbol which is not necessarily supposed to be exported,
make the driver boolean to avoid the build failure. Also, strengthen
the I2C dependency, which must also be built into the kernel for the
driver to compile.

BUG=chrome-os-partner:60540
TEST=Build arm:allmodconfig

Change-Id: I0eafa033455212d3f258f135f66c8bad05314597
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/416379
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@google.com>