CHROMIUM: kernel device_jail
It turns out we need to do this interposition in the kernel instead
of in another userspace process unless we want to basically re-
implement the device driver in userspace anyway, since ioctls are
likely to muck with userspace memory arbitrarily.

This registers a control device that you can make requests to in
order to create jails, and the jails will interpose IO syscalls.
Requests to open the jailed devices will be sent to whatever is
listening on /dev/jail-request. During normal use this should
be permission_broker.

BUG=chromium:644338
TEST=Open a crosh with a nexus player connected as usb device 1-2:
  # device_jail_utility --add=/dev/bus/usb/001/002
  # od -tx1 /dev/jailed-189-1
  0000000 12 01 10 02 00 00 00 40 d1 18 e7 4e ff ff 02 03
  0000020 04 01 09 02 25 00 01 01 00 80 fa 05 09 00 00 02
  0000040 09 04 00 00 02 ff 42 01 05 07 05 02 02 00 02 00
  0000060 07 05 82 02 00 02 00
  0000067
  device_jail_utility is found in CL:419188 and its server
  functionality can be used to mock out permission_broker as well.
  Running an adb container with the nexus player connected shows
  the device connected and I can access it with a shell using
  device_jail_fs mounted over /dev in the container.
  Trying to jail a non-USB device fails and logs to syslog that
  this is currently unsupported.

Original-reviewed-on: https://chromium-review.googlesource.com/418004
Original-reviewed-by: Dylan Reid <dgreid@chromium.org>
Change-Id: I00ca5ad62656614d91140cf5984caaa67b83a030
Signed-off-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/437912
Reviewed-by: Dylan Reid <dgreid@chromium.org>