FROMLIST: Bluetooth: btusb: Configure Marvell to use one of the pins for oob wakeup
The Marvell devices may have many gpio pins, and hence for wakeup
on these out-of-band pins, the chip needs to be told which pin is
to be used for wakeup, using an hci command.

Thus, we read the pin number etc from the device tree node and send
a command to the chip.

Signed-off-by: Rajat Jain <rajatja@google.com>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Acked-by: Rob Herring <robh@kernel.org>

(Cherry-picked from https://patchwork.kernel.org/patch/9537727/)

BUG=chrome-os-partner:57331
TEST=Verify wake-on-bt on kevin.

Change-Id: I5351b12fe09924126ebc0f1162e95a4f86d431d9
Reviewed-on: https://chromium-review.googlesource.com/419684
Commit-Ready: Rajat Jain <rajatja@chromium.org>
Tested-by: Rajat Jain <rajatja@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>