CHROMIUM: config: reduce mmap_rnd_bits from 32 to 31 for x86_64.
There seems to be a bug (either in the kernel or ASAN cryptohome) that causes
the range of ET_DYN to collide with the heap when 32 random bits are used.

Changing this to 31 would still provide plenty of randomness and allow us to
pass the relevant CTS tests. Meanwhile we will continue to investigate the
underlying problem.

BUG=chromium:674998
BUG=b:33398361
TEST=see instructions at http://crbug.com/674998#c27

Change-Id: I6137c5f3798e9de0ec6e57f9e4534d016ad72727
Reviewed-on: https://chromium-review.googlesource.com/426066
Commit-Ready: Haixia Shi <hshi@chromium.org>
Tested-by: Haixia Shi <hshi@chromium.org>
Reviewed-by: Haixia Shi <hshi@chromium.org>