FROMLIST: netfilter: ctnetlink: Fix regression in CTA_STATUS processing
The libnetfilter_conntrack userland library always sets IPS_CONFIRMED
when building a CTA_STATUS attribute.  If this toggles the bit from
0->1, the parser will return an error.  On Linux 4.4+ this will cause any
NFQA_EXP attribute in the packet to be ignored.  This breaks conntrackd's
userland helpers because they operate on unconfirmed connections.

Instead of returning -EBUSY if the user program asks to modify an
unchangeable bit, simply ignore the change.

Also, fix the logic so that user programs are allowed to clear
the bits that they are allowed to change.

Note that this changes CTA_STATUS parsing logic that has been in place
since ctnetlink was first introduced in 2.6.16.

BUG=b:29551145
TEST=manually run conntrackd on kevin, pyro

Link: https://patchwork.kernel.org/patch/9519973/
Fixes: b7bd1809e078 ("netfilter: nfnetlink_queue: get rid of nfnetlink_queue_ct.c")
Change-Id: Iea244ad64b02baeadc38e16d4d76e6273373b804
Signed-off-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/428493
Reviewed-by: Douglas Anderson <dianders@chromium.org>