CHROMIUM: drm/rockchip: Remove locks for dmcfreq handling
DMC frequency scaling needs to be bypassed if any of the following
conditions is true:
 1) more than 1 CRTC is active,
 2) any of the active CRTCs has a blanking period shorter than allowed by
   DMC frequency scaling.

This means that whenever we are changing any of the parameters affecting
the conditions above we also need to grab or release a dmcfreq block
reference. Even though code implementing this is already present, it
suffers from locking issues ultimately leading to deadlocks when
external monitor is connected, due to incorrect usage of modeset locks.

This patch rewrites the code to do necessary operations without said
locks by updating the state every time an atomic commit happens and
only for CRTCs affected by the commit, so that only commit state is
required to perform the dmcfreq condition update.

Note that the code that blocks/unblocks dmcfreq still locks a mutex.

BUG=chrome-os-partner:61875
TEST=Plug in an external monitor to kevin, see dmcfreq getting disabled,
  but no lockup.

Change-Id: Ie38f621e71cc362d621763241172d0a6197a1355
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/430397
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>