CHROMIUM: rk3399: ddrclk: Use smc to recalc rate
This changes the recalc rate function to use the smc call if the set
rate work is not busy. This is so the dmc devfreq driver can actually
tell if the set rate work timed out.

BUG=chrome-os-partner:62193
TEST=boot kevin
plug and unplug external monitor. Make sure that dpll and sclk_ddrc
are equal in /sys/kernel/debug/clk/clk_summary

Change-Id: I5eedb7bcf786f81c9631d3bf58bfa8db03d75871
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/431204
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>