FROMLIST: clk: rockchip: rk3288: make all niu clocks critical
NIU clocks are related to the interconnect and it's important to other blocks.
Since we don't have a driver to handle it, we should always enable it to avoid
casually close.

Make all of them critical,so that we don't have to each clock on its own
once things break.

Change-Id: I623bd4e2a45afadff65e33456cefdc717076f5c7
(cherry picked from commit 6eeec667316b0dd11923a5a8bff999a745407ae1)